requirements.txt
numpy
requests
flask
gunicorn
tensorflow-cpu